+++ auth.py
@@ -9,9 +9,10 @@ from ismobile import check
 class AuthLoginHandler(base.BaseHandler, tornado.auth.FacebookGraphMixin):
     @tornado.web.asynchronous
     def get(self):
+        next_url = self.get_safe_next_url("/")
         my_url = (self.request.protocol + "://" + self.request.host +
                   "/auth/login?next=" +
-                  tornado.escape.url_escape(self.get_argument("next", "/")))
+                  tornado.escape.url_escape(next_url))
         if self.get_argument("code", False):
             self.get_authenticated_user(redirect_uri=my_url, client_id=self.
                                         settings["facebook_api_key"], client_secret=self.settings[
@@ -28,9 +29,9 @@ class AuthLoginHandler(base.BaseHandler, tornado.auth.FacebookGraphMixin):
         #magic redirect to mobile if mobile or desktop if desktop
         user_agent = self.request.headers.get("User-Agent")
         if check(user_agent) == 1:
-            self.redirect(self.get_argument("next", "../events"))
+            self.redirect(self.get_safe_next_url("/events"))
         else:
-            self.redirect(self.get_argument("next", "../events"))
+            self.redirect(self.get_safe_next_url("/events"))
 
 class AuthLogoutHandler(base.BaseHandler, tornado.auth.FacebookGraphMixin):
     def get(self):